moved callbacks to callback directory, made single utils in model/
