+++ src/callbacks/__init__.py
@@ -1,0 +1,7 @@
+from .fid import FID
+from .save_generated_images import SaveGeneratedImages
+
+__all__ = [
+    "FID",
+    "SaveGeneratedImages",
+]
