+++ src/models/__init__.py
@@ -1,24 +1,12 @@
-from src.models.acgan import ACGAN
-from src.models.gan_utils import (
-    FID,
-    ConditionalDiscriminator,
-    ConditionalGenerator,
-    Discriminator,
-    Generator,
-    SaveGeneratedImages,
-    generate_and_save_images,
-)
-from src.models.vanilla_gan import VanillaGAN
+from src.models.acgan import ACGAN, ConditionalDiscriminator, ConditionalGenerator
+from src.models.vanilla_gan import Discriminator, Generator, VanillaGAN
 from src.models.wgan import WGAN_GP
 
 __all__ = [
-    "FID",
     "Discriminator",
     "Generator",
     "ConditionalGenerator",
     "ConditionalDiscriminator",
-    "SaveGeneratedImages",
-    "generate_and_save_images",
     "WGAN_GP",
     "ACGAN",
     "VanillaGAN",
