+++ src/models/vanilla_gan.py
@@ -8,10 +8,88 @@ import os
 from collections import defaultdict
 
 import torch
+import torch.nn as nn
 import torch.nn.functional as F
 from pytorch_lightning import LightningModule
 
-from .gan_utils import compute_metrics_no_aux, sample_image
+from .utils import compute_metrics_no_aux, sample_image, weights_init_normal
+
+
+class Generator(nn.Module):
+    def __init__(self, img_size: int, channels: int, latent_dim: int = 100):
+        super(Generator, self).__init__()
+        self.latent_dim = latent_dim
+        self.channels = channels
+
+        self.init_size = img_size // 4  # Initial size before upsampling
+        self.l1 = nn.Sequential(nn.Linear(latent_dim, 128 * self.init_size**2))
+
+        self.conv_blocks = nn.Sequential(
+            nn.BatchNorm2d(128),
+            nn.Upsample(scale_factor=2),
+            nn.Conv2d(128, 128, 3, stride=1, padding=1),
+            nn.BatchNorm2d(128, 0.8),
+            nn.LeakyReLU(0.2, inplace=True),
+            nn.Upsample(scale_factor=2),
+            nn.Conv2d(128, 64, 3, stride=1, padding=1),
+            nn.BatchNorm2d(64, 0.8),
+            nn.LeakyReLU(0.2, inplace=True),
+            nn.Conv2d(64, channels, 3, stride=1, padding=1),
+            nn.Tanh(),
+        )
+
+        # intialize with normal weights
+        self.apply(weights_init_normal)
+
+    def forward(self, noise):
+        out = self.l1(noise)
+        out = out.view(out.shape[0], 128, self.init_size, self.init_size)
+        img = self.conv_blocks(out)
+        return img
+
+
+class Discriminator(nn.Module):
+    def __init__(self, channels: int, img_size: int):
+        super(Discriminator, self).__init__()
+
+        def discriminator_block(in_filters, out_filters, bn=True):
+            """Returns layers of each discriminator block"""
+            block = [
+                nn.Conv2d(in_filters, out_filters, 3, 2, 1),
+                nn.LeakyReLU(0.2, inplace=True),
+                nn.Dropout2d(0.25),
+            ]
+            if bn:
+                block.append(nn.BatchNorm2d(out_filters, 0.8))
+            return block
+
+        self.conv_blocks = nn.Sequential(
+            *discriminator_block(channels, 16, bn=False),
+            *discriminator_block(16, 32),
+            *discriminator_block(32, 64),
+            *discriminator_block(64, 128),
+        )
+
+        # The height and width of downsampled image
+        if img_size == 28:
+            ds_size = 2
+        else:
+            # useful for imgs with size mod 2 == 0
+            ds_size = img_size // 2**4
+
+        # Output layers
+        self.adv_layer = nn.Sequential(nn.Linear(128 * ds_size**2, 1), nn.Sigmoid())
+
+        # intialize with normal weights
+        self.apply(weights_init_normal)
+
+    def forward(self, img):
+        out = self.conv_blocks(img)
+        out = out.view(out.shape[0], -1)
+        validity = self.adv_layer(out)
+
+        # output real/fake labels (validity)
+        return validity
 
 
 class VanillaGAN(LightningModule):
